moved datastore dir to nmmo/datastore
