+++ nmmo/datastore/id_allocator.py
